Docker-compose | Remove webflux dependencies
