+++ backend/bialydunajec-email/presentation/build.gradle.kts
@@ -18,8 +18,7 @@ dependencies {
     api(project(":bialydunajec-email:bialydunajec-email-messages"))
 
 
-    //compile("org.springframework.boot:spring-boot-starter-web")
-    compile("org.springframework.boot:spring-boot-starter-webflux")
+    compile("org.springframework.boot:spring-boot-starter-web")
     compile("com.fasterxml.jackson.module:jackson-module-kotlin")
 
     testCompile("org.springframework.boot:spring-boot-starter-test")
