+++ backend/bialydunajec-faq/presentation/build.gradle.kts
@@ -16,8 +16,7 @@ dependencies {
     api(project(":bialydunajec-ddd:bialydunajec-ddd-presentation"))
     api(project(":bialydunajec-faq:bialydunajec-faq-application"))
 
-    //compile("org.springframework.boot:spring-boot-starter-web")
-    compile("org.springframework.boot:spring-boot-starter-webflux")
+    compile("org.springframework.boot:spring-boot-starter-web")
     compile("com.fasterxml.jackson.module:jackson-module-kotlin")
 
     testCompile("org.springframework.boot:spring-boot-starter-test")
