+++ backend/bialydunajec-registrations/read-model/build.gradle.kts
@@ -18,5 +18,5 @@ dependencies {
     compile("org.springframework.boot:spring-boot-starter-data-mongodb")
     testCompile("org.springframework.boot:spring-boot-starter-test")
 
-    compile("org.springframework.boot:spring-boot-starter-webflux")
+    compile("org.springframework.boot:spring-boot-starter-web")
 }
